Add files via upload
For some reason, rgb_to_grayscale (and every other method to convert to monochrome) is producing images that apparently span a yellow-green-blue segment of the color spectrum.  Trying to resolve this; it's possible it's a simple matter of the default color map for plt.imshow(), and alternatively it's possible that for purposes of model building there's no real difference.  But still.  when I use a method called "rgb_to_grayscale" I would like to get an output that appears to be in grayscale.